refactor: enhance error handling in exercise actions for better reliability

+++ actions/exercises-action.ts
@@ -5,28 +5,40 @@ import type { ExerciseProps } from '@/types/data-types';
 import { revalidatePath } from 'next/cache';
 
 export const fetchExercises = async (): Promise<ExerciseProps[]> => {
-  return prisma.exercise.findMany({
-    orderBy: { createdAt: 'desc' },
-  });
+  try {
+    return prisma.exercise.findMany({
+      orderBy: { createdAt: 'desc' },
+    });
+  } catch (error) {
+    console.error('Error fetching exercises', error);
+    throw new Error(error instanceof Error ? error.message : 'Unexpected error');
+  }
 };
 
 export const createExercise = async (exerciseData: ExerciseProps): Promise<void> => {
   const { name, category, difficulty, imageUrl, description } = exerciseData;
 
-  await prisma.exercise.create({
-    data: {
-      name,
-      category,
-      difficulty,
-      imageUrl,
-      description,
-    },
-  });
+  try {
+    await prisma.exercise.create({
+      data: {
+        name,
+        category,
+        difficulty,
+        imageUrl,
+        description,
+      },
+    });
 
-  revalidatePath('/exercises');
+    revalidatePath('/exercises');
+  } catch (error) {
+    console.error('Error creating exercise:', error);
+    throw new Error(error instanceof Error ? error.message : 'Unexpected error');
+  }
 };
 
 export const deleteExercise = async (exerciseId: string): Promise<void> => {
+  if (!exerciseId) throw new Error('Missing ID for exercise deletion');
+
   try {
     await prisma.exercise.delete({
       where: {
@@ -37,12 +49,14 @@ export const deleteExercise = async (exerciseId: string): Promise<void> => {
     revalidatePath('/exercises');
   } catch (error) {
     console.error('Error deleting exercise:', error);
-    throw new Error('Failed to delete exercise');
+    throw new Error(error instanceof Error ? error.message : 'Unexpected error');
   }
 };
 
 export const updateExercise = async (exerciseData: ExerciseProps): Promise<void> => {
   const { name, category, difficulty, imageUrl, description, id } = exerciseData;
+
+  if (!id) throw new Error('Missing ID for exercise update');
 
   try {
     await prisma.exercise.update({
@@ -61,6 +75,6 @@ export const updateExercise = async (exerciseData: ExerciseProps): Promise<void>
     revalidatePath('/exercises');
   } catch (error) {
     console.error('Error updating exercise:', error);
-    throw new Error('Failed to update exercise');
+    throw new Error(error instanceof Error ? error.message : 'Unexpected error');
   }
 };
